fix(installation): exit of script when docker swarm init fails

+++ apps/website/public/feature.sh
@@ -45,6 +45,11 @@ install_dokploy() {
 
     docker swarm init --advertise-addr $advertise_addr
 
+    if [ $? -ne 0 ]; then
+        echo "Error: Failed to initialize Docker Swarm" >&2
+        exit 1
+    fi
+
     echo "Swarm initialized"
 
     docker network rm -f dokploy-network 2>/dev/null
